+++ src/domain/use_cases/countries/delete_country_use_case.ts
@@ -7,6 +7,9 @@ export class DeleteCountryUseCase {
 
   public async handle(id: number): Promise<null> {
     const country = await this.iCountryRepository.findById(id)
+    if (!country) {
+      throw new Error(`NotFound: tag with id ${id} not found`)
+    }
     return await this.iCountryRepository.delete(country)
   }
 }
