+++ src/domain/use_cases/stib_shape/delete_stib_use_case.ts
@@ -7,6 +7,9 @@ export class DeleteStibUseCase {
 
   public async handle(id: number): Promise<null> {
     const stibShape = await this.iStibRepository.findById(id)
+    if (!stibShape) {
+      throw new Error(`NotFound: Stib shape with id ${id} not found`)
+    }
     return await this.iStibRepository.delete(stibShape)
   }
 }
